notification-server/pyproject.toml
[tool.poetry]
name = "notification-server"
version = "0.1.0"
description = "Secure notification server receiving and displaying notifications from connected peers."
authors = ["Jan Kowalczyk <jan.p.kowalczyk@student.tugraz.at>"]

[tool.poetry.scripts]
notification-server = 'notification_server:main'

[tool.poetry.dependencies]
python = "^3.6"
py-notifier = "^0.1.0"
pystray = { git = "https://github.com/moses-palmer/pystray.git", branch = "master" }
ConfigArgParse = "^1.2.3"
PyYAML = "^5.3.1"
PyGObject = "^3.36.1"
pyopenssl-psk = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
